Move functions to Utils
