--- atcg_lib/platform/headless/src/Scene/ComponentGUIHandler.cpp
@@ -1,34 +1,0 @@
-#include <Scene/ComponentGUIHandler.h>
-
-namespace atcg
-{
-namespace GUI
-{
-
-AssetHandle displayMaterialSelection(const std::string& key, AssetHandle handle)
-{
-    return 0;
-}
-
-AssetHandle displayGraphSelection(const std::string& key, AssetHandle handle)
-{
-    return 0;
-}
-
-AssetHandle displayScriptSelection(const std::string& key, AssetHandle handle)
-{
-    return 0;
-}
-
-AssetHandle displayShaderSelection(const std::string& key, AssetHandle handle)
-{
-    return 0;
-}
-
-AssetHandle displayTexture2DSelection(const std::string& key, AssetHandle handle)
-{
-    return 0;
-}
-
-}    // namespace GUI
-}    // namespace atcg
