--- atcg_lib/src/Scene/ComponentSerializer.cpp
@@ -1,58 +1,0 @@
-#include <Scene/ComponentSerializer.h>
-
-#include <Scene/Entity.h>
-
-#include <DataStructure/TorchUtils.h>
-
-#include <Core/Path.h>
-
-namespace atcg
-{
-
-namespace Serialization
-{
-
-void serializeBuffer(const std::string& file_name, const char* data, const uint32_t byte_size)
-{
-    std::ofstream summary_file(file_name, std::ios::out | std::ios::binary);
-    summary_file.write(data, byte_size);
-    summary_file.close();
-}
-
-std::vector<uint8_t> deserializeBuffer(const std::string& file_name)
-{
-    std::ifstream summary_file(file_name, std::ios::in | std::ios::binary);
-    std::vector<uint8_t> buffer_char(std::istreambuf_iterator<char>(summary_file), {});
-    summary_file.close();
-
-    return buffer_char;
-}
-
-nlohmann::json serializeLayout(const atcg::BufferLayout& layout)
-{
-    nlohmann::json::array_t json_layout;
-    for(auto element: layout)
-    {
-        nlohmann::json::array_t json_element;
-        json_element.push_back((int)element.type);
-        json_element.push_back(element.name);
-
-        json_layout.push_back(json_element);
-    }
-
-    return json_layout;
-}
-
-atcg::BufferLayout deserializeLayout(nlohmann::json& layout_node)
-{
-    std::vector<atcg::BufferElement> elements;
-    for(nlohmann::json::array_t element: layout_node)
-    {
-        atcg::BufferElement buffer_element((atcg::ShaderDataType)element[0], element[1]);
-        elements.push_back(buffer_element);
-    }
-
-    return atcg::BufferLayout(elements);
-}
-}    // namespace Serialization
-}    // namespace atcg
