+++ atcg_lib/src/Scene/Components/CameraComponent.cpp
@@ -319,7 +319,7 @@ void ComponentGUIRenderer<CameraComponent>::draw_component(const atcg::ref_ptr<S
 
     ImGui::Separator();
 
-    auto new_handle = displayTexture2DSelection("camera", component.image_handle);
+    auto new_handle = Utils::displayTexture2DSelection("camera", component.image_handle);
 
     updated                = (new_handle != component.image_handle) || updated;
     component.image_handle = new_handle;
