+++ atcg_lib/src/Utils/Utils.cpp
@@ -291,5 +291,375 @@ setSkyLight(atcg::RendererSystem* renderer, const atcg::ref_ptr<Shader>& shader,
     return std::make_pair(irradiance_id, prefiltered_id);
 }
 
+AssetHandle displayMaterialSelection(const std::string& key, AssetHandle handle)
+{
+#ifndef ATCG_HEADLESS
+    const auto& data = AssetManager::getMetaData(handle);
+
+    std::string tag = AssetManager::isAssetHandleValid(handle) ? data.name : "Default Material";
+
+    const auto& registry = AssetManager::getAssetRegistry();
+
+    AssetHandle current_item = handle;
+
+    if(ImGui::BeginCombo(("Select Material##" + key).c_str(), tag.c_str()))
+    {
+        // No Selection
+        {
+            bool is_selected = !AssetManager::isAssetHandleValid(current_item);
+
+            if(ImGui::Selectable("Default Material", is_selected))
+            {
+                current_item = 0;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        for(auto it = registry.begin(); it != registry.end(); ++it)
+        {
+            if(it->second.type != AssetType::Material) continue;
+
+            bool is_selected = it->first == current_item;
+
+            if(ImGui::Selectable((it->second.name + "##" + std::to_string(it->first)).c_str(), is_selected))
+            {
+                current_item = it->first;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        ImGui::EndCombo();
+    }
+
+    return current_item;
+#else
+    return 0;
+#endif
+}
+
+AssetHandle displayGraphSelection(const std::string& key, AssetHandle handle)
+{
+#ifndef ATCG_HEADLESS
+    const auto& data = AssetManager::getMetaData(handle);
+
+    std::string tag = AssetManager::isAssetHandleValid(handle) ? data.name : "No Graph";
+
+    const auto& registry = AssetManager::getAssetRegistry();
+
+    AssetHandle current_item = handle;
+
+    if(ImGui::BeginCombo(("Select Graph##" + key).c_str(), tag.c_str()))
+    {
+        // No Selection
+        {
+            bool is_selected = !AssetManager::isAssetHandleValid(current_item);
+
+            if(ImGui::Selectable("No Graph", is_selected))
+            {
+                current_item = 0;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        for(auto it = registry.begin(); it != registry.end(); ++it)
+        {
+            if(it->second.type != AssetType::Graph) continue;
+
+            bool is_selected = it->first == current_item;
+
+            if(ImGui::Selectable((it->second.name + "##" + std::to_string(it->first)).c_str(), is_selected))
+            {
+                current_item = it->first;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        ImGui::EndCombo();
+    }
+
+    return current_item;
+#else
+    return 0;
+#endif
+}
+
+AssetHandle displayScriptSelection(const std::string& key, AssetHandle handle)
+{
+#ifndef ATCG_HEADLESS
+    const auto& data = AssetManager::getMetaData(handle);
+
+    std::string tag = AssetManager::isAssetHandleValid(handle) ? data.name : "No Script";
+
+    const auto& registry = AssetManager::getAssetRegistry();
+
+    AssetHandle current_item = handle;
+
+    if(ImGui::BeginCombo(("Select Script##" + key).c_str(), tag.c_str()))
+    {
+        // No Selection
+        {
+            bool is_selected = !AssetManager::isAssetHandleValid(current_item);
+
+            if(ImGui::Selectable("No Script", is_selected))
+            {
+                current_item = 0;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        for(auto it = registry.begin(); it != registry.end(); ++it)
+        {
+            if(it->second.type != AssetType::Script) continue;
+
+            bool is_selected = it->first == current_item;
+
+            if(ImGui::Selectable((it->second.name + "##" + std::to_string(it->first)).c_str(), is_selected))
+            {
+                current_item = it->first;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        ImGui::EndCombo();
+    }
+
+    return current_item;
+#else
+    return 0;
+#endif
+}
+
+AssetHandle displayShaderSelection(const std::string& key, AssetHandle handle)
+{
+#ifndef ATCG_HEADLESS
+    const auto& data = AssetManager::getMetaData(handle);
+
+    std::string tag = AssetManager::isAssetHandleValid(handle) ? data.name : "Default Shader";
+
+    const auto& registry = AssetManager::getAssetRegistry();
+
+    AssetHandle current_item = handle;
+
+    if(ImGui::BeginCombo(("Select Shader##" + key).c_str(), tag.c_str()))
+    {
+        // No Selection
+        {
+            bool is_selected = !AssetManager::isAssetHandleValid(current_item);
+
+            if(ImGui::Selectable("Default Shader", is_selected))
+            {
+                current_item = 0;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        for(auto it = registry.begin(); it != registry.end(); ++it)
+        {
+            if(it->second.type != AssetType::Shader) continue;
+
+            auto shader = AssetManager::getAsset<Shader>(it->first);
+            if(shader && shader->isComputeShader()) continue;
+
+            bool is_selected = it->first == current_item;
+
+            if(ImGui::Selectable((it->second.name + "##" + std::to_string(it->first)).c_str(), is_selected))
+            {
+                current_item = it->first;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        ImGui::EndCombo();
+    }
+
+    return current_item;
+#else
+    return 0;
+#endif
+}
+
+AssetHandle displayTexture2DSelection(const std::string& key, AssetHandle handle)
+{
+#ifndef ATCG_HEADLESS
+    const auto& data = AssetManager::getMetaData(handle);
+
+    std::string tag = AssetManager::isAssetHandleValid(handle) ? data.name : "No Image";
+
+    const auto& registry = AssetManager::getAssetRegistry();
+
+    AssetHandle current_item = handle;
+
+    if(ImGui::BeginCombo(("Select Image##" + key).c_str(), tag.c_str()))
+    {
+        // No Selection
+        {
+            bool is_selected = !AssetManager::isAssetHandleValid(current_item);
+
+            if(ImGui::Selectable("No Image", is_selected))
+            {
+                current_item = 0;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        for(auto it = registry.begin(); it != registry.end(); ++it)
+        {
+            if(it->second.type != AssetType::Texture2D) continue;
+
+            bool is_selected = it->first == current_item;
+
+            if(ImGui::Selectable((it->second.name + "##" + std::to_string(it->first)).c_str(), is_selected))
+            {
+                current_item = it->first;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        ImGui::EndCombo();
+    }
+
+    return current_item;
+#else
+    return 0;
+#endif
+}
+
+AssetHandle displayTexture3DSelection(const std::string& key, AssetHandle handle)
+{
+#ifndef ATCG_HEADLESS
+    const auto& data = AssetManager::getMetaData(handle);
+
+    std::string tag = AssetManager::isAssetHandleValid(handle) ? data.name : "No Image";
+
+    const auto& registry = AssetManager::getAssetRegistry();
+
+    AssetHandle current_item = handle;
+
+    if(ImGui::BeginCombo(("Select Image##" + key).c_str(), tag.c_str()))
+    {
+        // No Selection
+        {
+            bool is_selected = !AssetManager::isAssetHandleValid(current_item);
+
+            if(ImGui::Selectable("No Image", is_selected))
+            {
+                current_item = 0;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        for(auto it = registry.begin(); it != registry.end(); ++it)
+        {
+            if(it->second.type != AssetType::Texture3D) continue;
+
+            bool is_selected = it->first == current_item;
+
+            if(ImGui::Selectable((it->second.name + "##" + std::to_string(it->first)).c_str(), is_selected))
+            {
+                current_item = it->first;
+            }
+
+            if(is_selected)
+            {
+                ImGui::SetItemDefaultFocus();
+            }
+        }
+
+        ImGui::EndCombo();
+    }
+
+    return current_item;
+#else
+    return 0;
+#endif
+}
+
+void serializeBuffer(const std::string& file_name, const char* data, const uint32_t byte_size)
+{
+    std::ofstream summary_file(file_name, std::ios::out | std::ios::binary);
+    summary_file.write(data, byte_size);
+    summary_file.close();
+}
+
+std::vector<uint8_t> deserializeBuffer(const std::string& file_name)
+{
+    std::ifstream summary_file(file_name, std::ios::in | std::ios::binary);
+    std::vector<uint8_t> buffer_char(std::istreambuf_iterator<char>(summary_file), {});
+    summary_file.close();
+
+    return buffer_char;
+}
+
+nlohmann::json serializeLayout(const atcg::BufferLayout& layout)
+{
+    nlohmann::json::array_t json_layout;
+    for(auto element: layout)
+    {
+        nlohmann::json::array_t json_element;
+        json_element.push_back((int)element.type);
+        json_element.push_back(element.name);
+
+        json_layout.push_back(json_element);
+    }
+
+    return json_layout;
+}
+
+atcg::BufferLayout deserializeLayout(nlohmann::json& layout_node)
+{
+    std::vector<atcg::BufferElement> elements;
+    for(nlohmann::json::array_t element: layout_node)
+    {
+        atcg::BufferElement buffer_element((atcg::ShaderDataType)element[0], element[1]);
+        elements.push_back(buffer_element);
+    }
+
+    return atcg::BufferLayout(elements);
+}
+
 }    // namespace Utils
 }    // namespace atcg
